pyproject.toml
[project]
name = "madmax-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "markdown>=3.10",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "unstructured>=0.18.21",
]
